♻️refactor: AccountInfo,UserInfo DTO 추가
SOLID 원칙에 위배하던 코드들을 DTO를 분리함으로써 리팩토링 하였습니다
