+++ src/main/kotlin/Rainbow_Frends/domain/account/presentation/dto/AccountInfo.kt
@@ -1,0 +1,13 @@
+package Rainbow_Frends.domain.account.presentation.dto
+
+import Rainbow_Frends.domain.account.entity.Role
+
+data class AccountInfo(
+    val studentNum: Short,
+    val profilePictureName: String?,
+    val profilePictureUrl: String?,
+    val daramRole: Role?,
+    val lateNumber: Int?,
+    val roomNumber: Int?,
+    val floor: Int?
+)
