+++ src/main/kotlin/Rainbow_Frends/domain/account/service/AccountInfoService.kt
@@ -1,8 +1,9 @@
 package Rainbow_Frends.domain.account.service
 
-import Rainbow_Frends.domain.account.service.impl.AccountInfoServiceImpl
+import Rainbow_Frends.domain.account.presentation.dto.AccountInfo
+import Rainbow_Frends.domain.account.presentation.dto.UserInfo
 
 interface AccountInfoService {
-    fun getUserInfomation(email: String): AccountInfoServiceImpl.UserInfo
-    fun getAccountInfomation(grade: Byte, classNum: Byte, number: Byte): AccountInfoServiceImpl.AccountInfo?
+    fun getUserInfomation(email: String): UserInfo
+    fun getAccountInfomation(grade: Byte, classNum: Byte, number: Byte): AccountInfo?
 }
